자동번역기/requirements.txt
fastapi
uvicorn
openai
google-generativeai
python-dotenv
pandas
openpyxl
gunicorn
jinja2
python-multipart
